Disable resize and move of lifeline body
